server/requirements.txt
flask=3.0.0
flask_pymongo=2.3.0
python-dotenv=1.0.1
stripe=8.3.0